@@ -1,30 +1,36 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <OutputType>Exe</OutputType>
+        <IsPackable>false</IsPackable>
+        <TieredPGO>true</TieredPGO>        
+    </PropertyGroup>
 
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <NoWarn>1701;1702;RCS1090</NoWarn>
-  </PropertyGroup>
+    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+        <NoWarn>1701;1702;RCS1090</NoWarn>
+    </PropertyGroup>
 
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <NoWarn>1701;1702;RCS1090</NoWarn>
-  </PropertyGroup>
+    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+        <NoWarn>1701;1702;RCS1090</NoWarn>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="6.0.0-rc.1.21452.15" />
-  </ItemGroup>
+    <ItemGroup>
+        <None Remove="BenchmarkDotNet.Artifacts\**" />
+        <None Remove="run.cmd" />
+    </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\FastEndpointsBench\FastEndpointsBench.csproj" />
-    <ProjectReference Include="..\MapControllers\MapControllers.csproj" />
-    <ProjectReference Include="..\MinimalApi\MinimalApi.csproj" />
-    <ProjectReference Include="..\MvcControllers\MvcControllers.csproj" />
-  </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="BenchmarkDotNet" Version="0.13.5" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="7.0.4" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\FastEndpointsBench\FEBench.csproj" />
+        <ProjectReference Include="..\MinimalApi\MinimalApi.csproj" />
+        <ProjectReference Include="..\MvcControllers\MvcControllers.csproj" />
+    </ItemGroup>
 
 </Project>
