@@ -1,36 +1,33 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <GenerateDocumentationFile>true</GenerateDocumentationFile>
+        <NoWarn>CS1591;1701;1702;CA2016;RCS1090;CA2254;CS8618</NoWarn>
+    </PropertyGroup>
 
-  <PropertyGroup>
-    <NoWarn>1701;1702;CA2016</NoWarn>
-  </PropertyGroup>
+    <ItemGroup>
+      <Compile Remove="[Features]\Inventory\List\Search\**" />
+      <Content Remove="[Features]\Inventory\List\Search\**" />
+      <EmbeddedResource Remove="[Features]\Inventory\List\Search\**" />
+      <None Remove="[Features]\Inventory\List\Search\**" />
+    </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="5.0.10" />
-  </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="FluentValidation" Version="11.5.2" />
+    </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="[Features]\Customers\List\Search\" />
-    <Folder Include="[Features]\Inventory\List\Recent\" />
-    <Folder Include="[Features]\Inventory\List\Search\" />
-    <Folder Include="[Features]\Inventory\Manage\Delete\" />
-    <Folder Include="[Features]\Inventory\Manage\Update\" />
-    <Folder Include="[Features]\Sales\Orders\Create\" />
-    <Folder Include="[Features]\Sales\Orders\Retrieve\" />
-    <Folder Include="[Features]\Shipping\Dispatch\" />
-  </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\Src\Generator\FastEndpoints.Generator.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
+        <ProjectReference Include="..\Src\Library\FastEndpoints.csproj" />
+        <ProjectReference Include="..\Src\Security\FastEndpoints.Security.csproj" />
+        <ProjectReference Include="..\Src\Swagger\FastEndpoints.Swagger.csproj" />
+    </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Src\Library\FastEndpoints.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <InternalsVisibleTo Include="Test" />
-  </ItemGroup>
+    <ItemGroup>
+        <InternalsVisibleTo Include="Test" />
+    </ItemGroup>
 
 </Project>
