@@ -1,11 +1,10 @@
-﻿namespace Web.Auth
+﻿namespace Web.Auth;
+
+public static class Claim
 {
-    public static class Claim
-    {
-        public const string UserName = "username";
-        public const string AdminID = "admin-id";
-        public const string CustomerID = "customer-id";
-        public const string UserType = "user-type";
-        public const string NullClaim = "null-claim";
-    }
+    public const string UserName = "username";
+    public const string AdminID = "admin-id";
+    public const string CustomerID = "customer-id";
+    public const string UserType = "user-type";
+    public const string NullClaim = "null-claim";
 }
