@@ -1,9 +1,8 @@
-﻿namespace Web.Auth
+﻿namespace Web.Auth;
+
+public static class Role
 {
-    public static class Role
-    {
-        public const string Admin = "admin";
-        public const string Customer = "customer";
-        public const string Staff = "staff";
-    }
+    public const string Admin = "admin";
+    public const string Customer = "customer";
+    public const string Staff = "staff";
 }
