@@ -1,14 +1,17 @@
 ﻿using FastEndpoints;
+using FEBench;
 
 var builder = WebApplication.CreateBuilder();
+builder.Logging.ClearProviders();
 builder.Services.AddFastEndpoints();
+builder.Services.AddScoped<ScopedValidator>();
 
 var app = builder.Build();
 app.UseAuthorization();
 app.UseFastEndpoints();
 app.Run();
 
-namespace FastEndpointsBench
+namespace FEBench
 {
     public partial class Program { }
 }
