@@ -2,16 +2,13 @@ using FluentValidation;
 using MvcControllers;
 
 var builder = WebApplication.CreateBuilder(args);
-builder.Services
-    .AddSingleton<IValidator<Request>, Validator>()
-    .AddMvc();
+builder.Logging.ClearProviders();
+builder.Services.AddControllers();
+builder.Services.AddSingleton<IValidator<Request>, Validator>();
 
 var app = builder.Build();
-app.UseRouting();
 app.UseAuthorization();
-app.MapControllerRoute(
-    name: "default",
-    pattern: "{controller=Home}/{action=Index}/{id?}");
+app.MapControllers();
 app.Run();
 
 namespace MvcControllers
