@@ -1,19 +1,16 @@
-﻿using FastEndpoints;
+﻿namespace TestCases.MissingClaimTest;
 
-namespace TestCases.MissingClaimTest
+public class ThrowIfMissingEndpoint : Endpoint<ThrowIfMissingRequest>
 {
-    public class ThrowIfMissingEndpoint : Endpoint<ThrowIfMissingRequest>
+    public override void Configure()
     {
-        public ThrowIfMissingEndpoint()
-        {
-            Verbs(Http.POST);
-            Routes("/test-cases/missing-claim-test");
-        }
+        Verbs(Http.POST);
+        Routes("/test-cases/missing-claim-test");
+    }
 
-        protected override Task HandleAsync(ThrowIfMissingRequest req, CancellationToken ct)
-        {
-            //this line will never be reached as ErrorResponse will be sent due to claim missing
-            return Task.CompletedTask;
-        }
+    public override Task HandleAsync(ThrowIfMissingRequest req, CancellationToken ct)
+    {
+        //this line will never be reached as ErrorResponse will be sent due to claim missing
+        return Task.CompletedTask;
     }
 }
